+++ src/components/App.jsx
@@ -3,7 +3,8 @@ import { Route, Routes } from 'react-router-dom';
 import { Layout } from './Layout';
 
 import { ThemeProvider } from '@mui/material';
-import mainTheme from 'styles/theme';
+import mainTheme from 'styles/theme/theme';
+import Container from '@mui/material/Container';
 
 const HomePage = lazy(() => import('../pages/Home'));
 const DonatePage = lazy(() => import('../pages/Donate'));
@@ -13,14 +14,16 @@ export const App = () => {
   return (
     <>
       <ThemeProvider theme={mainTheme}>
-        <Routes>
-          <Route path="/" element={<Layout />}>
-            <Route index element={<HomePage />} />
-            <Route path="/donate" element={<DonatePage />} />
-            <Route path="*" element={<HomePage />} />
-            <Route path="/admin" element={<AdminPage />} />
-          </Route>
-        </Routes>
+        <Container>
+          <Routes>
+            <Route path="/" element={<Layout />}>
+              <Route index element={<HomePage />} />
+              <Route path="/donate" element={<DonatePage />} />
+              <Route path="*" element={<HomePage />} />
+              <Route path="/admin" element={<AdminPage />} />
+            </Route>
+          </Routes>
+        </Container>
       </ThemeProvider>
     </>
   );
